mission/requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.3.4
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cmake==3.18.4.post1
colorama==0.4.4
comtypes==1.1.7
cryptography==3.4.7
decorator==4.4.2
distlib==0.3.1
Django==3.0.10
django-bootstrap-themes==3.3.6
django-cities==0.6
django-crispy-forms==1.13.0
django-debug-toolbar==1.7
Django-Gtts==0.4
django-microsoft-auth==2.4.0
django-model-utils==4.1.1
django-mssql-backend==2.8.1
django-password-validation==0.1.1
django-rest-passwordreset==1.2.1
django-webpack-loader==0.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
dlib==19.22.0
docutils==0.16
ecdsa==0.14.1
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
future==0.18.2
gTTS==2.2.3
gTTS-token==1.1.4
idna==2.10
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
jellyfish==0.8.2
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
lib50==1.1.11
matplotlib-inline==0.1.2
ms-identity-web==0.16.6
msal==1.10.0
mssql-django==1.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.26
mysqlclient==2.0.3
networkx==2.5.1
numpy==1.20.3
oauthlib==3.1.1
opencv-python==4.5.2.52
parso==0.8.2
pefile==2019.4.18
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pipenv==2020.11.15
plumbum==1.7.0
pomegranate==0.14.5
prompt-toolkit==3.0.19
protobuf==3.17.3
psycopg2==2.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.20
pygame==2.0.1
Pygments==2.8.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
PyJWT==1.7.1
PyMySQL==1.0.2
pyodbc==4.0.32
pypiwin32==223
python-dateutil==2.8.1
python-jose==3.2.0
pyttsx3==2.90
pytube==10.8.5
pytz==2021.1
pywin32==300
pywin32-ctypes==0.2.0
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.25.1
requests-oauthlib==1.3.0
rpyc==5.0.1
rsa==4.7.2
scipy==1.6.3
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
submit50==3.0.0
swapper==1.1.2.post1
termcolor==1.1.0
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
urllib3==1.26.3
validators==0.18.2
virtualenv==20.4.2
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
